Guess The Number
The "Guess Number" game! is a mix of HTML, CSS, and JavaScript, this game asks you to guess a secret number between 1 and 20. The website gives you hints if your guess is too high or too low. It's easy to play on any device. Make guesses, get feedback, and work your way to uncovering the hidden number. It's a fun and engaging challenge that sharpens your thinking skills.